+++ web-common/src/features/sources/sourceUtils.ts
@@ -2,18 +2,9 @@ import type {
   V1ConnectorSpec,
   V1ReconcileError,
 } from "@rilldata/web-common/runtime-client";
-import { extractFileExtension } from "./extract-table-name";
-import type { SourceConnectionType } from "../../metrics/service/SourceEventTypes";
-import { behaviourEvent, errorEvent } from "../../metrics/initMetrics";
-import type { BehaviourEventMedium } from "../../metrics/service/BehaviourEventTypes";
-import type {
-  MetricsEventScreenName,
-  MetricsEventSpace,
-} from "../../metrics/service/MetricsTypes";
 import { getFilePathFromNameAndType } from "../entity-management/entity-mappers";
 import { EntityType } from "../entity-management/types";
-import { sanitizeEntityName } from "./extract-table-name";
-import { categorizeSourceError } from "./modal/errors";
+import { extractFileExtension, sanitizeEntityName } from "./extract-table-name";
 
 export function compileCreateSourceYAML(
   values: Record<string, unknown>,
@@ -126,45 +117,4 @@ export function getSourceError(errors: V1ReconcileError[], sourceName) {
   const path = getFilePathFromNameAndType(sourceName, EntityType.Table);
 
   return errors?.find((error) => error?.filePath === path);
-}
-
-export function emitSourceErrorTelemetry(
-  space: MetricsEventSpace,
-  screenName: MetricsEventScreenName,
-  errorMessage: string,
-  connectionType: SourceConnectionType,
-  fileName: string,
-) {
-  const categorizedError = categorizeSourceError(errorMessage);
-  const fileType = getFileTypeFromPath(fileName);
-  const isGlob = fileName.includes("*");
-
-  errorEvent?.fireSourceErrorEvent(
-    space,
-    screenName,
-    categorizedError,
-    connectionType,
-    fileType,
-    isGlob,
-  );
-}
-
-export function emitSourceSuccessTelemetry(
-  space: MetricsEventSpace,
-  screenName: MetricsEventScreenName,
-  medium: BehaviourEventMedium,
-  connectionType: SourceConnectionType,
-  fileName: string,
-) {
-  const fileType = getFileTypeFromPath(fileName);
-  const isGlob = fileName.includes("*");
-
-  behaviourEvent?.fireSourceSuccessEvent(
-    medium,
-    screenName,
-    space,
-    connectionType,
-    fileType,
-    isGlob,
-  );
 }
